% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_shading_polygon_function_sp.R
\name{hatched.SpatialPolygons}
\alias{hatched.SpatialPolygons}
\title{Return a SpatialLinesDataframe as hatched area for a SpatialPolygons}
\usage{
hatched.SpatialPolygons(x, density = 10, angle = 45, fillOddEven = FALSE)
}
\arguments{
\item{x}{SpatialPolygons* from library sp}

\item{density}{the density of shading lines, in lines per inch.  The
    default value of \code{NULL} means that no shading lines are drawn.
    A zero value of \code{density} means no shading nor filling whereas
    negative values and \code{NA} suppress shading (and so allow
    color filling).}

\item{angle}{the slope of shading lines, given as an angle
     in degrees (counter-clockwise).}

\item{fillOddEven}{logical controlling the polygon shading mode:  see
      below for details.  Default \code{FALSE}.}
}
\description{
Return a SpatialLinesDataframe as hatched area for a SpatialPolygons
}
