% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_shading_polygon_function_sp.R
\name{polygon.onehatch}
\alias{polygon.onehatch}
\title{Create one line for hatch area of one polygon}
\usage{
polygon.onehatch(x, y, x0, y0, xd, yd, ..debug.hatch = FALSE,
  fillOddEven = FALSE, ...)
}
\arguments{
\item{x}{vectors containing the coordinates of the vertices
    of the polygon.}

\item{y}{vectors containing the coordinates of the vertices
    of the polygon.}

\item{x0}{parameter as issued from \code{\link{polygon.fullhatch}}}

\item{y0}{parameter as issued from \code{\link{polygon.fullhatch}}}

\item{xd}{parameter as issued from \code{\link{polygon.fullhatch}}}

\item{yd}{parameter as issued from \code{\link{polygon.fullhatch}}}

\item{..debug.hatch}{for drawing when debugging function}

\item{fillOddEven}{logical controlling the polygon shading mode:  see
      below for details.  Default \code{FALSE}.}

\item{...}{graphical parameters such as \code{xpd}, \code{lend},
    \code{ljoin} and \code{lmitre} can be given as arguments.}
}
\description{
Create one line for hatch area of one polygon
}
